Fix/plane

+++ libs/libft/incs/libft/ft_nbrs.h
@@ -6,7 +6,7 @@
 /*   By: hsabir <marvin@42lausanne.ch>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/07 10:41:18 by hsabir            #+#    #+#             */
-/*   Updated: 2022/01/27 14:59:25 by hsabir           ###   ########.fr       */
+/*   Updated: 2022/02/15 17:09:37 by                  ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 typedef struct	s_atof
 {
-	double	int_part;
-	double	fract_part;
+	float	int_part;
+	float	fract_part;
 	int	div;
 	int	sign;
 	int	fraction;
@@ -24,6 +24,6 @@ typedef struct	s_atof
 
 char	*ft_itoa(int n);
 int		ft_atoi(const char *str);
-int		ft_atof(const char *str, double *n);
+int		ft_atof(const char *str, float *n);
 
 #endif
