+++ src/vector/rotate_vec.c
@@ -1,22 +1,22 @@
 #include <types.h>
 #include <math.h>
 
-t_vec3	rotate_y_axis(double angle, t_vec3 v)
+t_vec3	rotate_y_axis(float angle, t_vec3 v)
 {
 	t_vec3	r;
 
-	r.x = v.x * cos(angle) + v.z * sin(angle);
+	r.x = v.x * cosf(angle) + v.z * sinf(angle);
 	r.y = v.y;
-	r.z = -v.x * sin(angle) + v.z * cos(angle);
+	r.z = -v.x * sinf(angle) + v.z * cosf(angle);
 	return (r);
 }
 
-t_vec3	rotate_x_axis(double angle, t_vec3 v)
+t_vec3	rotate_x_axis(float angle, t_vec3 v)
 {
 	t_vec3	r;
 
 	r.x = v.x;
-	r.y = v.y * cos(angle) - v.z * sin(angle);
-	r.z = v.y * sin(angle) + v.z * cos(angle);
+	r.y = v.y * cosf(angle) - v.z * sinf(angle);
+	r.z = v.y * sinf(angle) + v.z * cosf(angle);
 	return (r);
 }
